SAK-14855
git-svn-id: https://source.sakaiproject.org/svn/sam/trunk@58392 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/samigo-services/src/java/org/sakaiproject/tool/assessment/facade/PublishedAssessmentFacadeQueries.java
@@ -2815,7 +2815,7 @@ public class PublishedAssessmentFacadeQueries extends HibernateDaoSupport
 	
 	private String fileSizeInKB(int fileSize) {
 		String fileSizeString = "1";
-		int size = Math.round(fileSize / 1024);
+		int size = Math.round((float)fileSize / 1024.0f);
 		if (size > 0) {
 			fileSizeString = size + "";
 		}
